+++ src/generators/AbstractGenerator.ts
@@ -1,6 +1,6 @@
 import { CommonInputModel, CommonModel, OutputModel, Preset, Presets } from '../models';
 import { InputProcessor } from '../processors';
-import { FormatHelpers, IndentationTypes } from '../helpers';
+import { IndentationTypes } from '../helpers';
 import { isPresetWithOptions } from '../utils';
 import { RenderOutput } from 'models/RenderOutput';
 
@@ -17,28 +17,6 @@ export const defaultGeneratorOptions: CommonGeneratorOptions = {
   indentation: {
     type: IndentationTypes.SPACES,
     size: 2,
-  }
-};
-
-/**
- * The common naming convention type shared between generators for different languages.
- */
-export type CommonNamingConvention = {
-  type?: (name: string | undefined, ctx: { model: CommonModel, inputModel: CommonInputModel }) => string;
-  property?: (test: string | undefined, ctx: { model: CommonModel, inputModel: CommonInputModel, property?: CommonModel }) => string;
-};
-
-/**
- * A CommonNamingConvention implementation shared between generators for different languages.
- */
-export const CommonNamingConventionImplementation: CommonNamingConvention = {
-  type: (name: string | undefined) => {
-    if (!name) {return '';}
-    return FormatHelpers.toPascalCase(name);
-  },
-  property: (test: string | undefined) => {
-    if (!test) {return '';}
-    return FormatHelpers.toCamelCase(test);
   }
 };
