+++ src/helpers/NameHelpers.ts
@@ -1,4 +1,5 @@
-import { CommonModel } from 'models';
+import { FormatHelpers } from '../helpers';
+import { CommonInputModel, CommonModel } from '../models';
 
 /**
  * Default property names for different aspects of the common model
@@ -21,3 +22,25 @@ export function getUniquePropertyName(rootModel: CommonModel, propertyName: stri
   }
   return propertyName;
 }
+
+/**
+ * The common naming convention type shared between generators for different languages.
+ */
+export type CommonNamingConvention = {
+  type?: (name: string | undefined, ctx: { model: CommonModel, inputModel: CommonInputModel }) => string;
+  property?: (test: string | undefined, ctx: { model: CommonModel, inputModel: CommonInputModel, property?: CommonModel }) => string;
+};
+
+/**
+ * A CommonNamingConvention implementation shared between generators for different languages.
+ */
+export const CommonNamingConventionImplementation: CommonNamingConvention = {
+  type: (name: string | undefined) => {
+    if (!name) {return '';}
+    return FormatHelpers.toPascalCase(name);
+  },
+  property: (test: string | undefined) => {
+    if (!test) {return '';}
+    return FormatHelpers.toCamelCase(test);
+  }
+};
