+++ test/generators/java/JavaRenderer.spec.ts
@@ -13,17 +13,16 @@ describe('JavaRenderer', () => {
   describe('nameType()', () => {
     test('should name the type', () => {
       const name = renderer.nameType('type__someType');
-      expect(name).toEqual('someType');
+      expect(name).toEqual('TypeSomeType');
     });
   });
   
   describe('nameProperty()', () => {
     test('should name the property', () => {
       const name = renderer.nameProperty('property__someProperty');
-      expect(name).toEqual('someProperty');
+      expect(name).toEqual('propertySomeProperty');
     });
   });
-
   describe('renderType()', () => {
     test('Should render refs with pascal case', () => {
       const model = new CommonModel();
